requirements.txt
mockito~=1.3.0
pybuilder~=0.13.3
setuptools~=57.0.0
pip~=21.3.1
wheel~=0.36.2
Flask~=2.0.2
six~=1.16.0
Jinja2~=3.0.3
psycopg2-binary~=2.9.3